% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{.extract_dataset_dictionaries}
\alias{.extract_dataset_dictionaries}
\title{Extract data dictionaries from week readmes}
\usage{
.extract_dataset_dictionaries(tt_data)
}
\arguments{
\item{tt_data}{A \code{tt_data} object from \code{\link[tidytuesdayR:tt_load]{tidytuesdayR::tt_load()}}.}
}
\value{
A named list of \code{tbl_df}s, with one tibble per dataset. The tibbles
have columns \code{variable} and \code{description}, and either one row per variable
or 0 rows.
}
\description{
Extract data dictionaries from week readmes
}
\keyword{internal}
