% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{.extract_tables}
\alias{.extract_tables}
\title{Extract all tables from a document}
\usage{
.extract_tables(x)
}
\arguments{
\item{x}{An xml (html) object or a subset thereof.}
}
\value{
A list of the tables in the object.
}
\description{
Extract all tables from a document
}
\keyword{internal}
