% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{.extract_text_cell}
\alias{.extract_text_cell}
\title{Extract text from a particular row of an html table row}
\usage{
.extract_text_cell(row, col_number)
}
\arguments{
\item{row}{An xml node_set of a single row.}

\item{col_number}{The 1-indexed column number to extract.}
}
\value{
A character scalar of the contents of that cell.
}
\description{
Extract text from a particular row of an html table row
}
\keyword{internal}
