% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{.extract_text_col}
\alias{.extract_text_col}
\title{Extract text from a column of an html table}
\usage{
.extract_text_col(rows, col_number)
}
\arguments{
\item{rows}{An xml node_set of rows.}

\item{col_number}{The 1-indexed column number to extract.}
}
\value{
A character vector of the contents of that column. Note that this is
\emph{only} the text (not, for example, urls).
}
\description{
Extract text from a column of an html table
}
\keyword{internal}
