% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{.extract_text_col_table}
\alias{.extract_text_col_table}
\title{Extract text from all rows of a column in a table}
\usage{
.extract_text_col_table(table, col_number)
}
\arguments{
\item{table}{An xml table object.}

\item{col_number}{The 1-indexed column number to extract.}
}
\value{
A character vector of the contents of that column. Note that this is
\emph{only} the text (not, for example, urls).
}
\description{
Extract text from all rows of a column in a table
}
\keyword{internal}
