% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{.extract_urls_cell}
\alias{.extract_urls_cell}
\title{Extract urls from a particular row of an html table row}
\usage{
.extract_urls_cell(row, col_number)
}
\arguments{
\item{row}{An xml node_set of a single row.}

\item{col_number}{The 1-indexed column number to extract.}
}
\value{
A character vector of any
}
\description{
Extract urls from a particular row of an html table row
}
\keyword{internal}
