% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{.extract_urls_col}
\alias{.extract_urls_col}
\title{Extract URLs from a column of an html table}
\usage{
.extract_urls_col(rows, col_number)
}
\arguments{
\item{rows}{An xml node_set of rows.}

\item{col_number}{The 1-indexed column number to extract.}
}
\value{
A list of character vectors of the urls in the table, with one vector
per row.
}
\description{
Extract URLs from a column of an html table
}
\keyword{internal}
