% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_tbl.R
\name{.parse_tt_year_readme}
\alias{.parse_tt_year_readme}
\title{Extract the weekly summary table from a TidyTuesday readme}
\usage{
.parse_tt_year_readme(year)
}
\arguments{
\item{year}{The year to extract.}
}
\value{
A table of information about TidyTuesday weeks for that year.
}
\description{
Extract the weekly summary table from a TidyTuesday readme
}
\keyword{internal}
