% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{.summarize_tt_dataset}
\alias{.summarize_tt_dataset}
\title{Summarize a TidyTuesday dataset}
\usage{
.summarize_tt_dataset(tt_dataset, year, week)
}
\arguments{
\item{tt_dataset}{The dataset to summarize.}

\item{year}{The year of this dataset (to include in the summary info).}

\item{week}{The week of this dataset (to include in the summary info).}
}
\value{
A \code{tbl_df} summarizing the data.
}
\description{
Summarize a TidyTuesday dataset
}
\keyword{internal}
