% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{.table_headings}
\alias{.table_headings}
\title{Extract headings from a table}
\usage{
.table_headings(table)
}
\arguments{
\item{table}{An xml/html table.}
}
\value{
A character vector of headings.
}
\description{
Extract headings from a table
}
\keyword{internal}
