% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{.tt_readme_rows}
\alias{.tt_readme_rows}
\title{Get dataset rows for a particular TidyTuesday year readme}
\usage{
.tt_readme_rows(year)
}
\arguments{
\item{year}{The year of the desired README.}
}
\value{
An xml node_set of rows.
}
\description{
Get dataset rows for a particular TidyTuesday year readme
}
\keyword{internal}
