% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{.tt_year_readme_table}
\alias{.tt_year_readme_table}
\title{Get a table from a particular TidyTuesday year readme}
\usage{
.tt_year_readme_table(year, table_number)
}
\arguments{
\item{year}{The year of the desired README.}

\item{table_number}{Which table on the page to load. I do not currently
validate this, so errors might be confusing if you send in a number higher
than the number of tables in that README.}
}
\value{
An xml node_set containing the body of the table.
}
\description{
Get a table from a particular TidyTuesday year readme
}
\keyword{internal}
