% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-metadata.R
\name{get_tt_datasets_metadata}
\alias{get_tt_datasets_metadata}
\title{Load and parse metadata about TidyTuesday datasets}
\usage{
get_tt_datasets_metadata(tt_tbl)
}
\arguments{
\item{tt_tbl}{A \code{tt_tbl} table returned by \code{\link[=get_tt_tbl]{get_tt_tbl()}}.}
}
\value{
A tbl_df with columns year, week, dataset_name, variables (number of
columns), observations (number of rows), and variable_details (names and
classes of all variables).
}
\description{
Load data week-by-week for all weeks in a tt_week_summaries, and parse the
datasets for information such as number of variables and number of
observations.
}
\seealso{
\code{\link{tt_datasets_metadata}} the resulting dataset.
}
