% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_tt_datasets_table}
\alias{get_tt_datasets_table}
\title{Load the Table of DataSets}
\usage{
get_tt_datasets_table(min_year = 2018L, max_year = this_year())
}
\arguments{
\item{min_year}{The minimum year to include, between 2018 and now.}

\item{max_year}{The maximum year to include, between 2018 and now.}
}
\value{
A tibble of information about TidyTuesday datasets in the requested
timeframe.
}
\description{
TidyTuesday has published a new dataset weekly since April 2018. This
function get the table of information about the dataset from each of those
weeks.
}
