% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_tbl.R
\name{get_tt_tbl}
\alias{get_tt_tbl}
\title{Load the TidyTuesday week summaries}
\usage{
get_tt_tbl(min_year = 2018L, max_year = this_year())
}
\arguments{
\item{min_year}{The minimum year to include, between 2018 and now.}

\item{max_year}{The maximum year to include, between 2018 and now.}
}
\value{
A \code{tt_tbl} tibble of information about TidyTuesday weeks in the
requested timeframe.
}
\description{
TidyTuesday has published a new dataset weekly since April 2018. This
function gets the table of information about each of those weeks.
}
