% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_tt_week_metadata}
\alias{get_tt_week_metadata}
\title{Load and Parse Metadata About a TidyTuesday Week}
\usage{
get_tt_week_metadata(year, week)
}
\arguments{
\item{year}{The year of the dataset.}

\item{week}{The integer week of that dataset within that year.}
}
\value{
A tbl_df with columns year, week, dataset_name, variables (number of
columns), observations (number of rows), and variable_details (names and
classes of all variables).
}
\description{
Load a week's TidyTuesday dataset(s), and calculate metadata.
}
