% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{parse_tt_urls}
\alias{parse_tt_urls}
\title{Create a URL dataset}
\usage{
parse_tt_urls(tt_tbl)
}
\arguments{
\item{tt_tbl}{A \code{tt_tbl} table returned by \code{\link[=get_tt_tbl]{get_tt_tbl()}}.}
}
\value{
A tbl_df with columns \code{year}, \code{week}, \code{type}, \code{url}, \code{scheme},
\code{domain}, \code{subdomain}, \code{tld}, \code{path}, \code{query}, and \code{fragment}.
}
\description{
Parse the \code{source_urls} and \code{article_urls} columns of a \code{tt_tbl} (returned by
\code{\link[=get_tt_tbl]{get_tt_tbl()}}) into a searchable url tibble.
}
