% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{this_year}
\alias{this_year}
\title{Extract the year portion of today's date}
\usage{
this_year()
}
\value{
The current year as an integer.
}
\description{
Extract the year of today's date in your computer's system timezone.
}
\examples{
this_year()
}
