% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-metadata.R
\docType{data}
\name{tt_datasets_metadata}
\alias{tt_datasets_metadata}
\title{TidyTuesday dataset metadata}
\format{
A data frame with one row per dataset (555 rows as of 2023-06-05) and
6 variables:
\describe{
\item{year (integer)}{The year in which the dataset was realeased.}
\item{week (integer)}{The week number for this dataset within this year.}
\item{dataset_name (character)}{The name of this dataset. Some weeks
have multiple datasets.}
\item{variable (integer)}{The number of columns in this dataset.}
\item{observations (integer)}{The number of rows in this dataset.}
\item{variable_details (tbl_df)}{A tibble with a row for each variable, and
6 columns:
\describe{
\item{variable (character)}{The name of this variable.}
\item{class (character)}{the class of this variable.}
\item{n_unique (integer)}{the number of unique values of this variable within this dataset.}
\item{min (list)}{The "lowest" value of this variable (lowest number, first value alphabetically, etc).}
\item{max (list)}{The "highest" value of this variable (highest number, last value alphabetically, etc).}
\item{description (character)}{A short description of this variable.}
}
}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday}
}
\usage{
tt_datasets_metadata
}
\description{
Metadata about the weekly TidyTuesday datasets.
}
\seealso{
\code{\link[=get_tt_datasets_metadata]{get_tt_datasets_metadata()}} for the function that is used to
compile this dataset.
}
\keyword{datasets}
