% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttapi-package.R
\docType{package}
\name{ttapi-package}
\alias{ttapi}
\alias{ttapi-package}
\title{ttapi: TidyTuesday Dataset API}
\description{
Find TidyTuesday datasets by variable class, and likely other things.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/ttapi}
  \item \url{https://r4ds.github.io/ttapi/}
  \item Report bugs at \url{https://github.com/r4ds/ttapi/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
