% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_tt_datasets_metadata}
\alias{get_tt_datasets_metadata}
\title{Load and Parse Metadata About a TidyTuesday DataSets}
\usage{
get_tt_datasets_metadata(tt_datasets_table)
}
\arguments{
\item{tt_datasets_table}{A \code{tt_datasets} table returned by
\code{\link[=get_tt_datasets_table]{get_tt_datasets_table()}}.}
}
\value{
A tbl_df with columns year, week, dataset_name, variables (number of
columns), observations (number of rows), and variable_details (names and
classes of all variables).
}
\description{
Load data week-by-week for all weeks in a tt_datasets_table, and parse the
datasets for information such as number of variables and number of
observations.
}
