% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_tt_week_metadata}
\alias{get_tt_week_metadata}
\title{Load and parse metadata about a TidyTuesday week}
\usage{
get_tt_week_metadata(year, week)
}
\arguments{
\item{year}{The year of the dataset.}

\item{week}{The integer week of that dataset within that year. Note that this
is the TidyTuesday week number, not necessarily the week number of that
year.}
}
\value{
A tbl_df with columns year, week, dataset_name, variables (number of
columns), observations (number of rows), and variable_details (names,
classes, n_unique (number of unique observations), min (or first
alphabetically) and max (or last alphabetically) of all variables).
}
\description{
Load a week's TidyTuesday dataset(s), and calculate metadata.
}
