% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttapi-package.R
\docType{package}
\name{ttmeta-package}
\alias{ttmeta}
\alias{ttmeta-package}
\title{ttmeta: TidyTuesday Dataset Metadata}
\description{
TidyTuesday has been releasing weekly datasets since 2018. This package provides metadata about those datasets, and functions to update that metadata.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/ttmeta}
  \item \url{https://r4ds.github.io/ttmeta/}
  \item Report bugs at \url{https://github.com/r4ds/ttmeta/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
